+++ xmfcn-spring-cloud-wechat-service/src/main/java/com/cn/xmf/service/music/service/MusicService.java
@@ -48,7 +48,7 @@ public class MusicService  {
           {
              return null;
           }
-          long totalcount =musicHelperService.getTotalCount(json);
+          int totalcount =musicHelperService.getTotalCount(json);
           List<Music> list= null;
           if(totalcount>0)
           {
